Prevent glass slabs from becoming transparent on sides/bottom.
For all glasslike drawtype nodes, the tiles contain up to 2
textures. The second texture is for "connected" glass, but that
likely is a transparent texture. If we use the `tiles` def without
modification, we will make all those slabs/slopes transparent on all
sides except the top.

To fix, we remove the tiles[2] from the tiledef in case the node is
a glasslike drawtype.