Add compressed dirt, remove redundant tar
Tar already exists in building_blocks, and the recipe to cook it comes into conflict with what's in streetsmod, where gravel can be cooked into asphalt. However, the texture for it is retained in case it needs to be derived from again in the future.
Dirt can be collected in as much quanitites as cobblestone can be, so in place of tar is a way to pack dirt when so much of it is collected.
The compressed dirt shares most properities with the compressed cobble, except that can be dug by hand, but digging it takes as long as with gravel. Its texture is a recolor of this mod's tar block.
In addition, a function for dirt sound has been added, and compressed dirt makes use of it.
The new node is inspired in part by a counterpart on the Xanadu server, where there is also a counterpart to compressed cobble.