Readd tar block. Fixes #103
The proper way to resolve the redundancy between `building_blocks:Tar` and `moreblocks:tar` is with an alias in building_blocks as noted in issue #68.

The recipe conflict with streetsmod has been resolved by using `default:pine_tree` in place of `default:gravel`.  This may also be more realistic than gravel as a source of tar because pine derived tar was once of major economic importance.