Improved placement for slabs etc
Stairs+ nodes now check which side of the face (top/bottom for
horizontal, left/right for vertical placement) is being clicked to
decide on the orientation of the placed node. The aux key can be used
to place the node as if from the opposite side.

When the under node and placed node are the same category (e.g. slab to
slab) then by default the orientation of the under node is copied,
flippingit if placing above or below an upright or upside-down node. The
aux key can be used to place node according to the general (no aux, no
same-cat) rules.

(And yes, this means that it's not possible to place a stair+ node
against another stair+ node as if it were from the other side. Sorry, we
don't have enough special keys.)