amend load print
Add a final **print to log** at the end of `init.lua` to indicate themod was loaded successfully. This idea was derived from other mods which are specifically aimed at MT servers.

This could be useful not only for player support (e.g. in the MT forum) but would be certainly useful for MT server admins using the CLI and reading their server logs.